gitlab-ci: switch to debian container images
Partly this is to make the CI look more like Mesa's, but also Debian's
infrastructure for cross-building is much less painful than Fedora's and
I'd like to get test coverage on non-x86.

We build an image that can cross-build for many architectures, but to
start we only build for amd64 and i386 since those are native and
therefore complete in a reasonable amount of time. We also do a ubsan
build for amd64, which is currently allowed to fail since we are not yet
ubsan-clean.

Other architectures (disabled in the yml) currently do not work as
written; they will find the OpenMP libraries at link time, but not at
runtime. Building them with OpenMP disabled would work, but the runtime
is then on the order of 20 minutes, which seems a bit too high.